Reorganize error logic
